signout and callback url fixed

+++ lib/acions/user.action.tsx
@@ -45,8 +45,9 @@ export async function signInWithCredentials(
 export async function signOutUser() {
   // get current users cart and delete it so it does not persist to next user
   const currentCart = await getMyCart()
-  await prisma.cart.delete({ where: { id: currentCart?.id } })
-  await signOut()
+
+  if (currentCart) await prisma.cart.delete({ where: { id: currentCart?.id } })
+  await signOut({ redirectTo: "/" })
 }
 
 export async function signUpUser(prevState: unknown, formData: FormData) {
